AirflowDemo/api/requirements.txt
Flask==2.0.1
PyYAML==5.4.1
requests==2.25.1
Werkzeug==2.0.1
